Rework package structure and remove some functions
- Fix issues with check_correct and check_or that were introduced in previous commit
- Drop check_df() and extends()
- Divide functions into check_funcs, has_funcs and check_logic files
- Move all old compound statement tests in the same file
- Update check_wrappers logic accordingly
- Update docs accordingly

Closes #297
