+++ pythonwhat/parsing.py
@@ -783,8 +783,6 @@ class WithParser(Parser):
             for item in items]
 
         tv_all = TargetVars(sum([list(c['target_vars'].items()) for c in context], []))
-        #body_tv = []
-        #for c in context: body_tv.extend(c['target_vars'])
 
         self.out.append({
             "context": context,
