--- pythonwhat/test_funcs/test_or.py
@@ -1,23 +1,0 @@
-from pythonwhat.Reporter import Reporter
-from pythonwhat.Test import TestFail, Test
-from pythonwhat.check_funcs import multi
-
-def check_or(*tests, state=None):
-    """Test whether at least one SCT passes."""
-
-    rep = Reporter.active_reporter
-
-    success = False
-    first_feedback = None
-    for test in tests: 
-        try: 
-            multi(test, state=state)
-            success = True
-        except TestFail as e:
-            if not first_feedback: first_feedback = e.feedback
-        if success: 
-            return
-    
-    rep.do_test(Test(first_feedback))
-
-test_or = check_or
